+++ crates/sncast/tests/e2e/deploy.rs
@@ -2,6 +2,7 @@ use crate::helpers::constants::{
     ACCOUNT, ACCOUNT_FILE_PATH, CONSTRUCTOR_WITH_PARAMS_CONTRACT_CLASS_HASH_SEPOLIA,
     DEVNET_OZ_CLASS_HASH_CAIRO_0, MAP_CONTRACT_CLASS_HASH_SEPOLIA, URL,
 };
+use crate::helpers::fee::apply_test_resource_bounds_flags;
 use crate::helpers::fixtures::{
     create_and_deploy_account, create_and_deploy_oz_account, get_transaction_hash,
     get_transaction_receipt,
@@ -11,7 +12,8 @@ use indoc::indoc;
 use shared::test_utils::output_assert::{assert_stderr_contains, assert_stdout_contains};
 use snapbox::cmd::{Command, cargo_bin};
 use sncast::AccountType;
-use sncast::helpers::constants::{ARGENT_CLASS_HASH, BRAAVOS_CLASS_HASH, OZ_CLASS_HASH};
+use sncast::helpers::constants::OZ_CLASS_HASH;
+use sncast::helpers::fee::FeeArgs;
 use starknet::core::types::TransactionReceipt::Deploy;
 use starknet_types_core::felt::Felt;
 use std::path::PathBuf;
@@ -34,9 +36,8 @@ async fn test_happy_case_human_readable() {
         "--salt",
         "0x2",
         "--unique",
-        "--max-fee",
-        "99999999999999999",
     ];
+    let args = apply_test_resource_bounds_flags(args);
 
     let snapbox = runner(&args).current_dir(tempdir.path());
     let output = snapbox.assert().success();
@@ -59,8 +60,9 @@ async fn test_happy_case_human_readable() {
 
 #[test_case(DEVNET_OZ_CLASS_HASH_CAIRO_0.parse().unwrap(), AccountType::OpenZeppelin; "cairo_0_class_hash")]
 #[test_case(OZ_CLASS_HASH, AccountType::OpenZeppelin; "cairo_1_class_hash")]
-#[test_case(ARGENT_CLASS_HASH, AccountType::Argent; "argent_class_hash")]
-#[test_case(BRAAVOS_CLASS_HASH, AccountType::Braavos; "braavos_class_hash")]
+#[test_case(sncast::helpers::constants::ARGENT_CLASS_HASH, AccountType::Argent; "argent_class_hash")]
+// TODO(#3089)
+// #[test_case(sncast::helpers::constants::BRAAVOS_CLASS_HASH, AccountType::Braavos; "braavos_class_hash")]
 #[tokio::test]
 async fn test_happy_case(class_hash: Felt, account_type: AccountType) {
     let tempdir = create_and_deploy_account(class_hash, account_type).await;
@@ -79,9 +81,8 @@ async fn test_happy_case(class_hash: Felt, account_type: AccountType) {
         "--salt",
         "0x2",
         "--unique",
-        "--max-fee",
-        "99999999999999999",
     ];
+    let args = apply_test_resource_bounds_flags(args);
 
     let snapbox = runner(&args).current_dir(tempdir.path());
 
@@ -91,19 +92,27 @@ async fn test_happy_case(class_hash: Felt, account_type: AccountType) {
     assert!(matches!(receipt, Deploy(_)));
 }
 
-#[test_case(Some("99999999999999999"), None, None; "max_fee")]
-#[test_case(None, Some("999"), None; "max_gas")]
-#[test_case(None, None, Some("999999999999"); "max_gas_unit_price")]
-#[test_case(None, None, None; "none")]
-#[test_case(Some("999999999999999"), None, Some("999999999999"); "max_fee_max_gas_unit_price")]
-#[test_case(None, Some("999"), Some("999999999999"); "max_gas_max_gas_unit_price")]
-#[test_case(Some("999999999999999"), Some("999"), None; "max_fee_max_gas")]
+// TODO(#3100)
+// #[test_case(FeeArgs{
+//     max_fee: Some(NonZeroFelt::try_from(Felt::from(1000000000000000000000000)).unwrap()),
+//     l1_data_gas: None,
+//     l1_data_gas_price:  None,
+//     l1_gas:  None,
+//     l1_gas_price:  None,
+//     l2_gas:  None,
+//     l2_gas_price:  None,
+// }; "max_fee")]
+#[test_case(FeeArgs{
+    max_fee: None,
+    l1_data_gas: Some(100_000),
+    l1_data_gas_price: Some(10_000_000_000_000),
+    l1_gas: Some(100_000),
+    l1_gas_price: Some(10_000_000_000_000),
+    l2_gas: Some(1_000_000_000),
+    l2_gas_price: Some(100_000_000_000_000_000_000),
+}; "resource_bounds")]
 #[tokio::test]
-async fn test_happy_case_different_fees(
-    max_fee: Option<&str>,
-    max_gas: Option<&str>,
-    max_gas_unit_price: Option<&str>,
-) {
+async fn test_happy_case_different_fees(fee_args: FeeArgs) {
     let tempdir = create_and_deploy_oz_account().await;
 
     let mut args = vec![
@@ -123,14 +132,31 @@ async fn test_happy_case_different_fees(
         "--unique",
     ];
     let options = [
-        ("--max-fee", max_fee),
-        ("--max-gas", max_gas),
-        ("--max-gas-unit-price", max_gas_unit_price),
+        (
+            "--max-fee",
+            fee_args.max_fee.map(Felt::from).map(|x| x.to_string()),
+        ),
+        ("--l1-data-gas", fee_args.l1_data_gas.map(|x| x.to_string())),
+        (
+            "--l1-data-gas-price",
+            fee_args.l1_data_gas_price.map(|x| x.to_string()),
+        ),
+        ("--l1-gas", fee_args.l1_gas.map(|x| x.to_string())),
+        (
+            "--l1-gas-price",
+            fee_args.l1_gas_price.map(|x| x.to_string()),
+        ),
+        ("--l2-gas", fee_args.l2_gas.map(|x| x.to_string())),
+        (
+            "--l2-gas-price",
+            fee_args.l2_gas_price.map(|x| x.to_string()),
+        ),
     ];
 
-    for &(key, value) in &options {
-        if let Some(val) = value {
-            args.append(&mut vec![key, val]);
+    for &(key, ref value) in &options {
+        if let Some(val) = value.as_ref() {
+            args.push(key);
+            args.push(val);
         }
     }
 
@@ -161,6 +187,7 @@ async fn test_happy_case_with_constructor() {
         "--class-hash",
         CONSTRUCTOR_WITH_PARAMS_CONTRACT_CLASS_HASH_SEPOLIA,
     ];
+    let args = apply_test_resource_bounds_flags(args);
 
     let snapbox = runner(&args);
     let output = snapbox.assert().success().get_output().stdout.clone();
@@ -190,6 +217,7 @@ async fn test_happy_case_with_constructor_cairo_expression_calldata() {
         "--class-hash",
         CONSTRUCTOR_WITH_PARAMS_CONTRACT_CLASS_HASH_SEPOLIA,
     ];
+    let args = apply_test_resource_bounds_flags(args);
 
     let snapbox = runner(&args).current_dir(tempdir.path());
     let output = snapbox.assert().success().get_output().stdout.clone();
@@ -200,6 +228,8 @@ async fn test_happy_case_with_constructor_cairo_expression_calldata() {
     assert!(matches!(receipt, Deploy(_)));
 }
 
+// TODO(#3116): Before, this test returned message 'Input too long for arguments').
+// Now, it returns message about transaction execution error.
 #[test]
 fn test_wrong_calldata() {
     let args = vec![
@@ -223,7 +253,7 @@ fn test_wrong_calldata() {
         output,
         indoc! {r"
         command: deploy
-        error: [..]('Input too long for arguments')[..]
+        error: Transaction execution error [..]
         "},
     );
 }
@@ -251,6 +281,8 @@ async fn test_contract_not_declared() {
     );
 }
 
+// TODO(#3116): Before, this test returned message containing info that contract is already deployed.
+// Now, it returns message about transaction execution error.
 #[test]
 fn test_contract_already_deployed() {
     let args = vec![
@@ -274,7 +306,7 @@ fn test_contract_already_deployed() {
         output,
         indoc! {r"
         command: deploy
-        error: [..]Requested ContractAddress[..]is unavailable for deployment[..]
+        error: Transaction execution error [..]
         "},
     );
 }
@@ -295,9 +327,11 @@ fn test_too_low_gas() {
         "--salt",
         "0x2",
         "--unique",
-        "--max-gas-unit-price",
+        "--l1-gas",
         "1",
-        "--max-gas",
+        "--l2-gas",
+        "1",
+        "--l1-data-gas",
         "1",
     ];
 
@@ -308,7 +342,7 @@ fn test_too_low_gas() {
         output,
         indoc! {r"
         command: deploy
-        error: Max fee is smaller than the minimal transaction cost
+        error: The transaction's resources don't cover validation or the minimal transaction fee
         "},
     );
 }