+++ scripts/install_devnet.sh
@@ -2,8 +2,9 @@
 set -e
 
 DEVNET_INSTALL_DIR="$(git rev-parse --show-toplevel)/crates/sncast/tests/utils/devnet"
-DEVNET_REPO="https://github.com/0xSpaceShard/starknet-devnet-rs"
-DEVNET_VER="v0.2.2"
+# TODO(#3087): Use official devnet release
+DEVNET_REPO="https://github.com/franciszekjob/starknet-devnet"
+DEVNET_VER="v0.13.4-rc.1"
 
 echo "Devnet installation directory: $DEVNET_INSTALL_DIR"